add require for pathname
