+++ spec/spec_helper.rb
@@ -4,6 +4,8 @@ require 'bundler/setup'
 require 'packs'
 require 'pry'
 
+require 'packs/rspec/support'
+
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
   config.example_status_persistence_file_path = '.rspec_status'
@@ -14,24 +16,4 @@ RSpec.configure do |config|
   config.expect_with :rspec do |c|
     c.syntax = :expect
   end
-
-  config.before do
-    Packs.bust_cache!
-  end
-
-  config.around do |example|
-    prefix = [File.basename($0), Process.pid].join('-') # rubocop:disable Style/SpecialGlobalVars
-    tmpdir = Dir.mktmpdir(prefix)
-    Dir.chdir(tmpdir) do
-      example.run
-    end
-  ensure
-    FileUtils.rm_rf(tmpdir)
-  end
-end
-
-def write_file(path, content = '')
-  pathname = Pathname.pwd.join(path)
-  FileUtils.mkdir_p(pathname.dirname)
-  pathname.write(content)
 end
